Fix intro image scaling if used in compound content types
When using QS in compount content types (in Column in particular),
the intro div may not have been rendered when determining its
width. In consequence, instead of scaling the intro image to full
width, it would be displayed in its natural dimensions in Column.

This pull request fixes this behavior and the intro image will scale
properly in Column as well.